+++ includes/modules/payment/easyinvoice.php
@@ -38,10 +38,10 @@
   
       if ($this->enabled === true) {
         if ((int) MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_ID > 0) {
-          $this->order_status = MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_ID;
+          $this->order_status = (int) MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_ID;
         }
         if ((int) MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_SUCCESS_ID > 0) {
-          $this->order_status_success = MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_SUCCESS_ID;
+          $this->order_status_success = (int) MODULE_PAYMENT_EASYINVOICE_ORDER_STATUS_SUCCESS_ID;
         }
         
         TeambankPayment::init($this->code);
